feat: project_detail 디자인 수정
- 뒤로가기 버튼 없앰(디자인이 깔끔해짐)
- description 실시간 업데이트 반영
- description 100자 & 10줄로 제한
- projectId는 project_detail에서 복사 가능
- 발표 시간 임시 삭제
- project_card와 work_dialogs 파일이 바뀐 것 해결
