+++ app/lib/src/pages/work_page/work_page.dart
@@ -1,13 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import 'package:talk_pilot/src/pages/work_page/widgets/project_detail_page.dart';
-import 'package:talk_pilot/src/pages/work_page/widgets/work_dialogs.dart';
-import 'package:talk_pilot/src/pages/work_page/widgets/project_card.dart';
-
+import 'package:talk_pilot/src/models/project_model.dart';
 import 'package:talk_pilot/src/provider/user_provider.dart';
 import 'package:talk_pilot/src/provider/project_provider.dart';
-import 'package:talk_pilot/src/models/project_model.dart';
+
+import 'package:talk_pilot/src/pages/work_page/widgets/work_dialogs.dart';
+import 'package:talk_pilot/src/pages/work_page/widgets/project_card.dart';
+import 'package:talk_pilot/src/pages/work_page/widgets/project_detail_page.dart';
 
 class WorkPage extends StatefulWidget {
   const WorkPage({super.key});
